split up logic more accordingly. hashing for various httpclients should be handled as a self contained method, and lru cache http client should make no opinion about this. TODO: add tests for changes to RequestsHttpClient
