+++ test/http/test_lru_http_client.py
@@ -1,9 +1,34 @@
 from lru_cache_http_client.hash.hasher import Hasher
-from lru_cache_http_client.http.requests_http_client import HttpClient
+from lru_cache_http_client.http.http_client import HttpClient
 from lru_cache_http_client.http.lru_http_client import LruHttpClient
 
 import pytest
 import time
+from collections.abc import Hashable
+
+
+class TestHttpClient(HttpClient):
+    """
+    Mock HttpClient to be injected into LruHttpClient for testing
+    Unique calls to `get` will return unique vals.
+    This is useful as we're trying to test the caching abilities
+    of LruHttpClient.
+    """
+
+    count = 1
+
+    def get(self, url, params=None, **kwargs):
+        self.count += 1
+        return self.count
+
+    def make_args_hashable(self, params=None, **kwargs):
+        """
+        Hashability tests are outside of the scope of the test suite
+        """
+        for key, val in kwargs.items():
+            if not isinstance(val, Hashable):
+                kwargs[key] = None
+        return None, kwargs
 
 
 def test_invalid_http_client():
@@ -41,13 +66,6 @@ def test_same_req():
            different results
     Assert: Both responses are the same due to caching
     """
-
-    class TestHttpClient(HttpClient):
-        count = 1
-
-        def get(self, url, params=None, **kwargs):
-            self.count += 1
-            return self.count
 
     client = TestHttpClient()
     caching_client = LruHttpClient(http_client=client)
@@ -97,14 +115,6 @@ def test_req_url_params_list():
     Assert: LruHttpClient is able to make the list hashable
             and caches the request
     """
-
-    class TestHttpClient(HttpClient):
-        count = 1
-
-        def get(self, url, params=None, **kwargs):
-            self.count += 1
-            return self.count
-
     client = TestHttpClient()
     caching_client = LruHttpClient(http_client=client)
     url_params = [("a", "b"), ("c", "d")]
@@ -122,13 +132,6 @@ def test_ttl_expired():
     Assert: The injected HttpClient's get method is called twice
     """
 
-    class TestHttpClient(HttpClient):
-        counter = 0
-
-        def get(self, url, params=None, **kwargs):
-            self.counter += 1
-            return self.counter
-
     class TestHasher(Hasher):
         counter = 0
 
@@ -141,5 +144,5 @@ def test_ttl_expired():
     caching_client = LruHttpClient(http_client=client, hasher=hasher)
     res1 = caching_client.get("abc")
     res2 = caching_client.get("abc")
-    assert res1 == 1
-    assert res2 == 2
+    assert res1 == 2
+    assert res2 == 3
